backend/requirements.txt
pandas
ollama
lizard
plotly
bs4
fastapi
uvicorn
sqlalchemy
psycopg2
celery
redis
langchain
langchain_community
langchain-ollama
jsonschema
pydantic
python-dotenv